pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robo-pokemon"
version = "0.1.0"
description = "Template de automacao de games com arquitetura em camadas e plugin por jogo."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.1",
  "pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
robo-game = "app.cli:main"
